+++ src/background/telemetry/index.ts
@@ -15,6 +15,7 @@ export {
     type SidebarLinkItemClickActionNames,
     type DnsServerClickActionNames,
     type FreeGbItemClickActionNames,
+    type LocationsTabClickActionNames,
 } from './telemetryEnums';
 
 export const telemetry = new Telemetry({