+++ src/options/components/Exclusions/Search/index.ts
@@ -1,3 +1,2 @@
 export { ExclusionsSearch } from './ExclusionsSearch';
 export { ServicesSearch } from './ServicesSearch';
-export { SearchHighlighter } from './SearchHighlighter';