+++ src/popup/components/Locations/Locations.tsx
@@ -1,18 +1,26 @@
-import React, { useContext, useRef } from 'react';
+import React, { useContext, useRef, useState } from 'react';
 import { observer } from 'mobx-react';
 
 import { rootStore } from '../../stores';
 import { reactTranslator } from '../../../common/reactTranslator';
+import { translator } from '../../../common/translator';
 import { type LocationData } from '../../stores/VpnStore';
 import { useTelemetryPageViewEvent } from '../../../common/telemetry';
 import { TelemetryActionName, TelemetryScreenName } from '../../../background/telemetry';
+import { Icon } from '../ui/Icon';
 
 import { FastestSkeleton } from './FastestSkeleton';
 import { Location } from './Location';
 import { Search } from './Search';
 import { Reload } from './Reload';
+import { TabButtons } from './TabButtons';
 
 import './endpoints.pcss';
+
+/**
+ * Duration of the notification that appears when a location is deleted in milliseconds.
+ */
+const DELETED_NOTIFICATION_DURATION_MS = 3000;
 
 export const Locations = observer(() => {
     const {
@@ -24,10 +32,21 @@ export const Locations = observer(() => {
 
     const isSearchTelemetrySent = useRef(false);
 
+    const deletedNotificationTimeout = useRef<NodeJS.Timeout>();
+    const [lastUnsavedLocation, setLastUnsavedLocation] = useState<string | null>(null);
+
     useTelemetryPageViewEvent(
         telemetryStore,
         TelemetryScreenName.LocationsScreen,
     );
+
+    const {
+        filteredLocations,
+        fastestLocationsToDisplay,
+        showSearchResults,
+        isSavedLocationsTab,
+        notSearchingAndSavedTab,
+    } = vpnStore;
 
     const handleLocationSelect = async (id: string) => {
         const prevId = vpnStore.selectedLocation?.id;
@@ -46,6 +65,31 @@ export const Locations = observer(() => {
         }
     };
 
+    const handleLocationSave = async (id: string) => {
+        const isAdded = await vpnStore.toggleSavedLocation(id);
+
+        if (!isAdded) {
+            clearTimeout(deletedNotificationTimeout.current);
+            setLastUnsavedLocation(id);
+
+            deletedNotificationTimeout.current = setTimeout(() => {
+                setLastUnsavedLocation(null);
+            }, DELETED_NOTIFICATION_DURATION_MS);
+        }
+    };
+
+    const handleNotificationUndo = () => {
+        if (lastUnsavedLocation) {
+            vpnStore.addSavedLocation(lastUnsavedLocation);
+            setLastUnsavedLocation(null);
+        }
+    };
+
+    const handleNotificationClose = () => {
+        setLastUnsavedLocation(null);
+        clearTimeout(deletedNotificationTimeout.current);
+    };
+
     const handleLocationsClose = () => {
         uiStore.closeEndpointsSearch();
         vpnStore.setSearchValue('');
@@ -55,8 +99,10 @@ export const Locations = observer(() => {
         return (
             <Location
                 key={location.id}
-                handleClick={handleLocationSelect}
                 location={location}
+                searchValue={vpnStore.searchValue}
+                onClick={handleLocationSelect}
+                onSaveClick={handleLocationSave}
             />
         );
     });
@@ -83,24 +129,25 @@ export const Locations = observer(() => {
     };
 
     const renderFilteredEndpoint = () => {
-        const {
-            filteredLocations,
-            showSearchResults,
-        } = vpnStore;
-
         const emptySearchResults = showSearchResults && filteredLocations.length === 0;
-        let listTitle = 'endpoints_all';
-
-        if (showSearchResults && filteredLocations.length > 0) {
-            listTitle = 'endpoints_search_results';
-        }
-
         if (emptySearchResults) {
             return (
                 <div className="endpoints__not-found">
-                    <div className="endpoints__not-found--icon" />
-                    <div className="endpoints__title endpoints__not-found--title">
-                        {reactTranslator.getMessage('endpoints_not_found')}
+                    <div className="endpoints__not-found-icon" />
+                    <div className="endpoints__title endpoints__not-found-title">
+                        {translator.getMessage('endpoints_not_found')}
+                    </div>
+                </div>
+            );
+        }
+
+        const emptySavedLocations = isSavedLocationsTab && filteredLocations.length === 0;
+        if (emptySavedLocations) {
+            return (
+                <div className="endpoints__empty-saved">
+                    <Icon icon="bookmark-off-thin" className="endpoints__empty-saved-icon" />
+                    <div className="endpoints__empty-saved-title">
+                        {reactTranslator.getMessage('endpoints_empty_saved')}
                     </div>
                 </div>
             );
@@ -108,37 +155,28 @@ export const Locations = observer(() => {
 
         return (
             <div className="endpoints__list">
-                <div className="endpoints__title">
-                    {reactTranslator.getMessage(listTitle)}
-                    {!showSearchResults && (
-                        <>
-                            &nbsp;
-                            {`(${filteredLocations.length})`}
-                        </>
-                    )}
-                </div>
+                {notSearchingAndSavedTab && (
+                    <div className="endpoints__title">
+                        {translator.getMessage('endpoints_all')}
+                        &nbsp;
+                        {`(${filteredLocations.length})`}
+                    </div>
+                )}
                 {renderLocations(filteredLocations)}
             </div>
         );
     };
 
-    const {
-        fastestLocationsToDisplay,
-        showSearchResults,
-    } = vpnStore;
-
     return (
         <div className="endpoints">
             <div className="endpoints__header">
-                {reactTranslator.getMessage('endpoints_countries')}
+                {translator.getMessage('endpoints_countries')}
                 <button
                     type="button"
                     className="button endpoints__back"
                     onClick={handleLocationsClose}
                 >
-                    <svg className="icon icon--button">
-                        <use xlinkHref="#back" />
-                    </svg>
+                    <Icon icon="back" className="icon--button" />
                 </button>
                 <Reload />
             </div>
@@ -147,11 +185,12 @@ export const Locations = observer(() => {
                 onChange={handleSearchInput}
                 onClear={handleSearchClear}
             />
+            {!showSearchResults && <TabButtons />}
             <div className="endpoints__scroll">
-                {!showSearchResults && (
+                {notSearchingAndSavedTab && (
                     <div className="endpoints__list">
                         <div className="endpoints__title">
-                            {reactTranslator.getMessage('endpoints_fastest')}
+                            {translator.getMessage('endpoints_fastest')}
                         </div>
                         {fastestLocationsToDisplay.length > 0 ? (
                             renderLocations(fastestLocationsToDisplay)
@@ -163,6 +202,39 @@ export const Locations = observer(() => {
 
                 {renderFilteredEndpoint()}
             </div>
+
+            {lastUnsavedLocation && (
+                <div className="endpoints__notification">
+                    <div className="endpoints__notification-wrapper">
+                        <Icon
+                            icon="info"
+                            className="endpoints__notification-icon"
+                        />
+                        <div className="endpoints__notification-content">
+                            <div className="endpoints__notification-title">
+                                {translator.getMessage('endpoints_saved_location_deleted')}
+                            </div>
+                            <button
+                                type="button"
+                                className="endpoints__notification-undo"
+                                onClick={handleNotificationUndo}
+                            >
+                                {translator.getMessage('settings_exclusions_undo')}
+                            </button>
+                        </div>
+                        <button
+                            type="button"
+                            className="endpoints__notification-close"
+                            onClick={handleNotificationClose}
+                        >
+                            <Icon
+                                icon="cross"
+                                className="endpoints__notification-close-icon icon--button"
+                            />
+                        </button>
+                    </div>
+                </div>
+            )}
         </div>
     );
 });