+++ src/popup/stores/GlobalStore.ts
@@ -78,11 +78,16 @@ export class GlobalStore {
                 shouldShowMobileEdgePromoBanner,
                 isVpnBlocked,
                 isHostPermissionsGranted,
+                locationsTab,
+                savedLocationIds,
             } = popupData;
 
             settingsStore.setForwarderDomain(forwarderDomain);
 
             telemetryStore.setIsHelpUsImproveEnabled(helpUsImprove);
+
+            vpnStore.setLocationsTab(locationsTab);
+            vpnStore.setSavedLocationIds(savedLocationIds);
 
             if (!isAuthenticated) {
                 await authStore.handleInitPolicyAgreement(policyAgreement);